Use conftest.py to share fixtures among tests, and begin removing duplicated fixtures from tests.
